pyproject.toml
[project]
name = "star-voting-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "py-cord[voice]>=2.6.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
